+++ entity/src/main/java/com/edaisong/entity/domain/OrderDetailBusiness.java
@@ -1,6 +1,6 @@
 package com.edaisong.entity.domain;
 
-import java.math.BigDecimal;
+import java.lang.Double;
 import java.util.Date;
 
 /**
@@ -23,10 +23,10 @@ public class OrderDetailBusiness {
 	private String receviceAddress;
 	private String trueName;
 	private String phoneNo;
-	private BigDecimal amount;
-	private BigDecimal orderCommission;
+	private Double amount;
+	private Double orderCommission;
 	private int isPay;
-	private BigDecimal settleMoney;
+	private Double settleMoney;
 	private Date cancelTime;
 
 	/**
@@ -228,28 +228,28 @@ public class OrderDetailBusiness {
 	/**
 	 * 金额
 	 */
-	public BigDecimal getAmount() {
+	public Double getAmount() {
 		return amount;
 	}
 
 	/**
 	 * 金额
 	 */
-	public void setAmount(BigDecimal amount) {
+	public void setAmount(Double amount) {
 		this.amount = amount;
 	}
 
 	/**
 	 * 佣金
 	 */
-	public BigDecimal getOrderCommission() {
+	public Double getOrderCommission() {
 		return orderCommission;
 	}
 
 	/**
 	 * 佣金
 	 */
-	public void setOrderCommission(BigDecimal orderCommission) {
+	public void setOrderCommission(Double orderCommission) {
 		this.orderCommission = orderCommission;
 	}
 
@@ -272,7 +272,7 @@ public class OrderDetailBusiness {
 	 * 
 	 * @return
 	 */
-	public BigDecimal getSettleMoney() {
+	public Double getSettleMoney() {
 		return settleMoney;
 	}
 
@@ -281,7 +281,7 @@ public class OrderDetailBusiness {
 	 * 
 	 * @param settleMoney
 	 */
-	public void setSettleMoney(BigDecimal settleMoney) {
+	public void setSettleMoney(Double settleMoney) {
 		this.settleMoney = settleMoney;
 	}
